feat(modal): split out modal components into their own files
