+++ packages/modal/src/Modal.tsx
@@ -1,45 +1,9 @@
-import type { Variants } from 'framer-motion';
-import type { MouseEvent, ReactNode } from 'react';
+import type { ReactNode } from 'react';
 
-import { ButtonGroup, IconButton } from '@launchpad-ui/button';
-import { FocusTrap } from '@launchpad-ui/focus-trap';
-import { Close, Warning } from '@launchpad-ui/icons';
-import { Portal } from '@launchpad-ui/portal';
-import { usePreventScroll } from '@react-aria/overlays';
-import { cx } from 'classix';
-import { LazyMotion, m } from 'framer-motion';
-import { useEffect, useRef } from 'react';
-
-import { MODAL_LABELLED_BY } from './constants';
-import styles from './styles/Modal.module.css';
-import { useMediaQuery } from './utils';
-
-const overlay: Variants = {
-  visible: { opacity: 1, transition: { duration: 0.15 } },
-  hidden: { opacity: 0 },
-};
-
-const transitions: { [key: string]: Variants } = {
-  desktopPop: {
-    hidden: { opacity: 0, scale: 0.9 },
-    visible: { opacity: 1, scale: 1, transition: { type: 'spring', delay: 0.1, duration: 0.15 } },
-  },
-  mobileSlideUp: {
-    hidden: { opacity: 0, y: '25%' },
-    visible: {
-      opacity: 1,
-      y: '0%',
-      transition: { type: 'spring', delay: 0.15, duration: 0.2, bounce: 0 },
-    },
-  },
-};
-
-const loadFeatures = () =>
-  import(
-    /* webpackChunkName: "lp-modal-framer-features" */
-    /* webpackExports: "domAnimation" */
-    'framer-motion'
-  ).then((res) => res.domAnimation);
+import { ModalBody } from './ModalBody';
+import { ModalContainer } from './ModalContainer';
+import { ModalFooter } from './ModalFooter';
+import { ModalHeader } from './ModalHeader';
 
 type ModalProps = {
   children: ReactNode;
@@ -65,7 +29,7 @@ const Modal = ({
   children,
   onReady,
   onCancel,
-  size = 'normal',
+  size,
   status,
   title,
   hasRequiredField,
@@ -74,127 +38,25 @@ const Modal = ({
   description,
   'data-test-id': testId = 'modal',
 }: ModalProps) => {
-  const ref = useRef<HTMLDivElement>(null);
-  const isDesktopViewport = useMediaQuery('(min-width: 430px)');
-
-  usePreventScroll();
-
-  useEffect(() => {
-    const handleEscape = (event: KeyboardEvent) => {
-      event.stopImmediatePropagation();
-      const latest = [...document.querySelectorAll('[data-modal]')].pop();
-      if (event.key === 'Escape' && latest === ref.current) {
-        close();
-      }
-    };
-
-    const close = () => {
-      onCancel?.();
-    };
-
-    onReady?.();
-    document.body.classList.add('has-overlay');
-    document.addEventListener('keydown', handleEscape);
-
-    return () => {
-      document.body.classList.remove('has-overlay');
-      document.removeEventListener('keydown', handleEscape);
-    };
-  }, [onReady, onCancel, testId]);
-
-  const handleOverlayClick = (event: MouseEvent<HTMLDivElement>) => {
-    if (cancelWithOverlayClick && event.target === event.currentTarget) {
-      onCancel && onCancel();
-    }
-  };
-
   const hasFooter = !!(primaryButton || secondaryButton);
 
   return (
-    <Portal>
-      <LazyMotion strict features={loadFeatures}>
-        <div
-          className={cx(styles.modal, styles[size], className)}
-          data-modal
-          data-test-id={testId}
-          ref={ref}
-        >
-          <m.div
-            initial="hidden"
-            animate="visible"
-            variants={overlay}
-            transition={{ duration: 0.15 }}
-            role="presentation"
-            className={styles.overlay}
-            data-test-id="modal-overlay"
-            onMouseDown={handleOverlayClick}
-          >
-            <FocusTrap autoFocus restoreFocus>
-              <m.div
-                initial="hidden"
-                animate="visible"
-                variants={isDesktopViewport ? transitions.desktopPop : transitions.mobileSlideUp}
-                role="dialog"
-                aria-labelledby={MODAL_LABELLED_BY}
-                aria-modal
-                className={styles.content}
-                tabIndex={-1}
-              >
-                <div data-test-id="modal-header" className={styles.header}>
-                  <div className={styles.headerMain}>
-                    {status === 'warning' && (
-                      <Warning
-                        data-test-id="modal-header-icon"
-                        size="medium"
-                        className={styles.headerIcon}
-                      />
-                    )}
-                    <h2 id={MODAL_LABELLED_BY} data-test-id="modal-title" className={styles.title}>
-                      {title}
-                    </h2>
-                    {withCloseButton && (
-                      <IconButton
-                        aria-label="close"
-                        icon={<Close size="medium" />}
-                        className={styles.closeButton}
-                        onClick={onCancel}
-                        data-test-id="modal-close-button"
-                      />
-                    )}
-                  </div>
-                  {description && (
-                    <p className={styles.headerDescription} data-test-id="modal-description">
-                      {description}
-                    </p>
-                  )}
-                  {hasRequiredField && (
-                    <div
-                      className={styles.headerRequiredFields}
-                      data-test-id="modal-required-field"
-                    >
-                      <span className={styles.requiredAsterisk}>*</span> Required field
-                    </div>
-                  )}
-                </div>
+    <ModalContainer
+      onCancel={onCancel}
+      onReady={onReady}
+      cancelWithOverlayClick={cancelWithOverlayClick}
+      size={size}
+      className={className}
+      data-test-id={testId}
+    >
+      <ModalHeader
+        {...{ withCloseButton, title, status, onCancel, description, hasRequiredField }}
+      />
 
-                <div className={styles.body} data-test-id="modal-body">
-                  {children}
-                </div>
+      <ModalBody>{children}</ModalBody>
 
-                {hasFooter && (
-                  <div className={styles.footer} data-test-id="modal-footer">
-                    <ButtonGroup className={styles.footerActions}>
-                      {secondaryButton}
-                      {primaryButton}
-                    </ButtonGroup>
-                  </div>
-                )}
-              </m.div>
-            </FocusTrap>
-          </m.div>
-        </div>
-      </LazyMotion>
-    </Portal>
+      {hasFooter && <ModalFooter secondaryButton={secondaryButton} primaryButton={primaryButton} />}
+    </ModalContainer>
   );
 };
